Move LSP constants to completion manager so provider depend on manager and not languageserver
