+++ spyder/plugins/completion/kite/providers/document.py
@@ -15,7 +15,7 @@ import os.path as osp
 
 from qtpy.QtCore import QMutexLocker
 from spyder.plugins.completion.kite.decorators import send_request, handles
-from spyder.plugins.completion.languageserver import (
+from spyder.plugins.completion.manager.api import (
     LSPRequestTypes, CompletionItemKind)
