+++ spyder/plugins/completion/languageserver/client.py
@@ -28,7 +28,7 @@ import psutil
 # Local imports
 from spyder.config.base import (DEV, get_conf_path, get_debug_level,
                                 running_under_pytest)
-from spyder.plugins.completion.languageserver import (
+from spyder.plugins.completion.manager.api import (
     CLIENT_CAPABILITES, SERVER_CAPABILITES,
     TEXT_DOCUMENT_SYNC_OPTIONS, LSPRequestTypes,
     ClientConstants)
