+++ spyder/plugins/completion/languageserver/plugin.py
@@ -25,9 +25,9 @@ from qtpy.QtWidgets import QMessageBox
 from spyder.config.base import _, get_conf_path, running_under_pytest
 from spyder.config.lsp import PYTHON_CONFIG
 from spyder.config.manager import CONF
-from spyder.api.completion import SpyderCompletionPlugin
 from spyder.utils.misc import check_connection_port, getcwd_or_home
-from spyder.plugins.completion.languageserver import LSP_LANGUAGES
+from spyder.plugins.completion.manager.api import (LSP_LANGUAGES,
+                                                   SpyderCompletionPlugin)
 from spyder.plugins.completion.languageserver.client import LSPClient
 from spyder.plugins.completion.languageserver.confpage import (
     LanguageServerConfigPage)
