+++ spyder/plugins/completion/languageserver/providers/client.py
@@ -8,7 +8,7 @@
 
 import logging
 
-from spyder.plugins.completion.languageserver import LSPRequestTypes
+from spyder.plugins.completion.manager.api import LSPRequestTypes
 from spyder.plugins.completion.languageserver.decorators import (
     handles, send_response)
