+++ spyder/plugins/completion/languageserver/providers/window.py
@@ -8,7 +8,7 @@
 
 import logging
 
-from spyder.plugins.completion.languageserver import LSPRequestTypes
+from spyder.plugins.completion.manager.api import LSPRequestTypes
 from spyder.plugins.completion.languageserver.decorators import handles
 
 logger = logging.getLogger(__name__)
