+++ spyder/plugins/completion/languageserver/tests/test_client.py
@@ -12,7 +12,7 @@ from qtpy.QtCore import QObject, Signal, Slot
 
 from spyder.config.lsp import PYTHON_CONFIG
 from spyder.plugins.completion.languageserver.client import LSPClient
-from spyder.plugins.completion.languageserver import (
+from spyder.plugins.completion.manager.api import (
     LSPRequestTypes, WorkspaceUpdateKind)
