+++ spyder/plugins/editor/widgets/completion.py
@@ -19,7 +19,7 @@ from qtpy.QtWidgets import (QAbstractItemView, QApplication, QListWidget,
 # Local imports
 from spyder.utils import icon_manager as ima
 from spyder.plugins.completion.kite.providers.document import KITE_COMPLETION
-from spyder.plugins.completion.languageserver import CompletionItemKind
+from spyder.plugins.completion.manager.api import CompletionItemKind
 from spyder.py3compat import to_text_string
 from spyder.widgets.helperwidgets import HTMLDelegate
